fixed_remarks
