+++ src/test/java/BaseTest.java
@@ -1,0 +1,38 @@
+import org.junit.After;
+import org.junit.Before;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import ru.yandex.praktikum.model.LoginPage;
+import ru.yandex.praktikum.model.MainPage;
+
+import java.time.Duration;
+
+public class BaseTest {
+
+    public WebDriver driver;
+    public MainPage mainPage;
+    public LoginPage loginPage;
+
+    @Before
+    public void setUp() {
+        ChromeOptions options = new ChromeOptions();
+        System.setProperty("webdriver.chrome.driver", "C:\\WebDriver\\bin\\yandexdriver.exe");
+        options.setBinary("C:\\Users\\1\\AppData\\Local\\Yandex\\YandexBrowser\\Application\\browser.exe");
+        driver = new ChromeDriver(options); // Для запуска тестов в яндекс браузере
+
+        //driver = new ChromeDriver(); // Для запуска тестов в хром браузере
+    }
+
+    @Before
+    public void transitionMainPage() {
+        mainPage = new MainPage(driver);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
+        mainPage.open();
+    }
+
+    @After
+    public void tearDown() {
+        driver.quit();
+    }
+}
