+++ src/test/java/LoginTest.java
@@ -1,38 +1,13 @@
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 import ru.yandex.praktikum.model.ForgotPasswordPage;
 import ru.yandex.praktikum.model.LoginPage;
-import ru.yandex.praktikum.model.MainPage;
 import ru.yandex.praktikum.model.RegistrationPage;
-
-import java.time.Duration;
 
 import static org.junit.Assert.assertEquals;
 
-public class LoginTest {
-
-    private WebDriver driver;
-    private MainPage mainPage;
-
-    @Before
-    public void setUp() {
-        ChromeOptions options = new ChromeOptions();
-        System.setProperty("webdriver.chrome.driver", "C:\\WebDriver\\bin\\yandexdriver.exe");
-        options.setBinary("C:\\Users\\1\\AppData\\Local\\Yandex\\YandexBrowser\\Application\\browser.exe");
-        driver = new ChromeDriver(options); // Для запуска тестов в яндекс браузере
-
-        //driver = new ChromeDriver(); // Для запуска тестов в хром браузере
-
-        mainPage = new MainPage(driver);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-        mainPage.open();
-    }
+public class LoginTest extends BaseTest {
 
     @Test
     @DisplayName("Авторизация пользователя через кнопку личного кабинета")
@@ -94,10 +69,5 @@ public class LoginTest {
         loginPage.passwordForUser();
         loginPage.clickForLoginUser();
         assertEquals("Соберите бургер", mainPage.isMessageTitle());
-    }
-
-    @After
-    public void cleanUp() {
-        driver.quit();
     }
 }
