+++ src/test/java/MainConstructorPage.java
@@ -1,36 +1,15 @@
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 import ru.yandex.praktikum.model.LoginPage;
-import ru.yandex.praktikum.model.MainPage;
 
-import java.time.Duration;
+import static org.junit.Assert.assertEquals;
 
-import static org.junit.Assert.*;
-
-public class MainConstructorPage {
-
-    private WebDriver driver;
-    private MainPage mainPage;
-    private LoginPage loginPage;
+public class MainConstructorPage extends BaseTest{
 
     @Before
-    public void setUp() {
-        ChromeOptions options = new ChromeOptions();
-        System.setProperty("webdriver.chrome.driver", "C:\\WebDriver\\bin\\yandexdriver.exe");
-        options.setBinary("C:\\Users\\1\\AppData\\Local\\Yandex\\YandexBrowser\\Application\\browser.exe");
-        driver = new ChromeDriver(options); // Для запуска тестов в яндекс браузере
-
-        //driver = new ChromeDriver(); // Для запуска тестов в хром браузере
-
-        mainPage = new MainPage(driver);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(5));
-        mainPage.open();
+    public void signIn() {
         mainPage.clickPersonalAccount();
 
         loginPage = new LoginPage(driver);
@@ -85,10 +64,5 @@ public class MainConstructorPage {
         mainPage.clickBuns();
 
         assertEquals("Булки", mainPage.changedClassConstructor());
-    }
-
-    @After
-    public void cleanUp() {
-        driver.quit();
     }
 }
